app/requirements.txt
streamlit==1.41.1
Pillow==10.0.1
pytesseract==0.3.13
opencv-python>=4.7.0,<5.0.0
#opencv-python==4.9.0
numpy==1.26.0
onnxruntime==1.18.1
scikit-image==0.18.1